Add silent mode and throw instead of exiting
The silent_mode option has been added to the SCAMPArgs struct. If this
option is set to true (false by default) the execution will not print to
the standard output.

The error handling method has been changed. The pointes where exit or
terminate were being called now throw an exception which is handled from
main.
The error checking done for allocation and deallocation in the
constructor and destructor of qt_compute_helper have been moved to
specific functions that are called at the begin and end of the qt
computation. It is necessary to avoid throwing from the destructor.